Fix to start P2PQuake.exe when update fails
